+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Pending
 
+* Renamed NewVideoOverlay -> VideoOverlay, new-video-overlay -> video-overlay [#8152](https://github.com/Applifier/unity-ads-webview/pull/8152)
+* Add close button to the new Privacy settigns view [#8106](https://github.com/Applifier/unity-ads-webview/pull/8106)
+
+# Fri Feb  8 20:37:59 UTC 2019
+
 * Install button experiment for rewarded ad videos [#8055](https://github.com/Applifier/unity-ads-webview/pull/8055)
+* Update dependencies [#8054](https://github.com/Applifier/unity-ads-webview/pull/8054)
 
 # Thu Feb  7 00:49:07 UTC 2019
 * Fix connect to disconnect in MRAIDAdapterContainer [#8150](https://github.com/Applifier/unity-ads-webview/pull/8150)